// frontend/src/api/model/fileTypeEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `log` - Log file
* `trajectory` - Trajectory
* `document` - Document
* `image` - Image
* `other` - Other
* `parquet` - Parquet
 */
export type FileTypeEnum = typeof FileTypeEnum[keyof typeof FileTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FileTypeEnum = {
  log: 'log',
  trajectory: 'trajectory',
  document: 'document',
  image: 'image',
  other: 'other',
  parquet: 'parquet',
} as const;