// frontend/src/api/model/fileUploadRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */
import type { FileTypeEnum } from './fileTypeEnum';

export interface FileUploadRequest {
  /**
   * @minLength 1
   * @maxLength 200
   */
  filename: string;
  file: Blob;
  file_type?: FileTypeEnum;
  metadata?: unknown;
  /** @nullable */
  owner?: string | null;
}